I had created a state to hold each cards id number instead of useing the value passed down because I thought it had to update. Switched it back to just reference the id that was passed down and now it works again. All requests are now handled at the top RelatedProducts module.
